tests/wpt/tests/css/css-grid/grid-lanes/tentative/parsing/grid-lanes-direction-computed.html

<!DOCTYPE html>
<html>
<head>
  <meta charset="utf-8">
  <title>CSS Grid Lanes: grid-lanes-direction getComputedStyle()</title>
  <link rel="author" title="Celeste Pan" href="mailto:celestepan@microsoft.com">
  <link rel="help" href="https://drafts.csswg.org/css-grid-3">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/css/support/computed-testcommon.js"></script>
  <script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
  <div id="target"></div>
  </div>
  <script>
    test_computed_value("grid-lanes-direction", "row");
    test_computed_value("grid-lanes-direction", "column");
    test_computed_value("grid-lanes-direction", "row track-reverse");
    test_computed_value("grid-lanes-direction", "row fill-reverse");
    test_computed_value("grid-lanes-direction", "row track-reverse fill-reverse", "row fill-reverse track-reverse");
    test_computed_value("grid-lanes-direction", "row fill-reverse track-reverse");
    test_computed_value("grid-lanes-direction", "column track-reverse");
    test_computed_value("grid-lanes-direction", "column fill-reverse");
    test_computed_value("grid-lanes-direction", "column track-reverse fill-reverse", "column fill-reverse track-reverse");
    test_computed_value("grid-lanes-direction", "column fill-reverse track-reverse");
    test_computed_value("grid-lanes-direction", "normal");
  </script>
</body>
</html>
